Conflicts:
	bot.py

+++ manage.py
@@ -20,6 +20,9 @@ while not good:
         good = False
 
 command = sys.argv[1]
+if (command == "list"):
+    os.system("""curl http://$ES_HOST:9200/_cat/indices?v""")
+    sys.exit(0)
 index = sys.argv[2]
 if (command == "delete"):
     os.system("curl -XDELETE http://$ES_HOST:9200/" + index)
